Add ability to define a custom hook in C#
To test it, you can use the function
UnityHaptics.SetHook(UnityHaptics.GravityHook).

The parameter of UnityHaptics.SetHook() is a function that takes 2
Vector3 (position and velocity) and returns a Vector3 (force). This
hook is run during each haptic loop (~1kHz, so the function called
must be fast) and the resulting force is added to the one computed by
the library itself (interaction forces).